FE-Afi-25: Edit Flow Frontend add to cart, product detail show in modal
